backend/foodgram/requirements.txt
Django==4.2.5
flake8==6.1.0
isort==5.12.
psycopg==3.1.10
pillow==10.0.1
djangorestframework==3.14.0